cloud-architecture-a2/requirements.txt
absl-py==0.9.0
aiokafka==0.7.0
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
autopep8==1.5.4
awscli==1.18.69
awscrt==0.6.2
AWSIoTPythonSDK==1.4.8
awsiotsdk==1.3.2
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==3.1.5
boto==2.49.0
boto3==1.12.46
botocore==1.15.49
bs4==0.0.1
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
click-plugins==1.1.1
cligj==0.5.0
cmake==3.17.3
colorama==0.4.3
confluent-kafka==1.5.0
contextlib2==0.6.0.post1
cryptography==2.9.2
cvdata==0.0.7
cycler==0.10.0
Cython==0.29.20
cytoolz==0.10.1
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
Delorean==1.0.0
descartes==1.1.0
detecto==1.1.4
docutils==0.15.2
easydict==1.9
EditorConfig==0.12.2
entrypoints==0.3
fabric==2.5.0
Fiona==1.8.13.post1
Flask==1.1.2
future==0.18.2
gast==0.3.3
geopandas==0.8.1
glob2==0.7
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hdbscan==0.8.26
humanize==3.0.1
idna==2.9
iiif-downloader==0.0.8
ImageHash==4.1.0
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
imutils==0.5.3
install==1.3.3
invoke==1.4.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.0
iso8601==0.1.13
itsdangerous==0.24
jedi==0.17.2
Jinja2==2.10.1
jmespath==0.9.5
joblib==0.16.0
jsbeautifier==1.13.0
json5==0.9.5
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.7
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
kafka-python==2.0.1
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
llvmlite==0.33.0
lxml==4.5.1
m3u8==0.7.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
munch==2.5.0
natsort==7.0.1
nbclient==0.5.0
nbconvert==6.0.2
nbformat==5.0.7
nest-asyncio==1.4.0
networkx==2.4
nltk==3.5
notebook==6.1.4
numba==0.50.1
numpy==1.19.1
oauthlib==3.1.0
onnx==1.7.0
onnxruntime==1.3.0
opencv-python==3.4.3.18
openimages==0.0.1
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.7.1
pascal-voc-writer==0.1.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pixplot==0.0.88
pointgrid==0.0.3
prettytable==0.7.2
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.12.0
psutil==5.7.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.0
pylab-sdk==1.1.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyproj==2.6.1.post1
PyQt5==5.15.0
PyQt5-sip==12.8.0
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-ffmpeg-video-streaming==0.1.5
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.10.11
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
schedule==0.6.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
SQLAlchemy-Paginator==0.2
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
tmdbsimple==2.6.6
toml==0.10.1
toolz==0.10.0
torch==1.6.0
torchfile==0.1.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
transitions==0.8.2
typing-extensions==3.7.4.2
tzlocal==2.1
umap-learn==0.4.6
urllib3==1.25.9
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.15.3
widgetsnbextension==3.5.1
wrapt==1.12.1
yale-dhlab-keras-preprocessing==1.1.1
yale-dhlab-rasterfairy==1.0.3
zipp==3.1.0
